refactor accommodation entity

+++ Restino.Application/Contracts/Persistance/IReservationRepository.cs
@@ -6,7 +6,7 @@ namespace Restino.Application.Contracts.Persistance
     {
         Task<bool> IsGuestsCountWithinCapacity(int guestsCount, Guid accommodationId);
         Task<bool> IsDateRangeValid(DateTime checkInDate, DateTime checkOutDate, Guid accommodationId);
-        Task<double> TotalPrice(Guid accommodationId, DateTime checkInDate, DateTime checkOutDate);
+        Task<decimal> TotalPrice(Guid accommodationId, DateTime checkInDate, DateTime checkOutDate);
         Task<List<Reservations>> ListUserReservations(string userId);
     }
 }
